milestone 2: conviction voting (#41)
* initiate for conviction voting

* implement vote recording

* veto and proposal cleaning

* correct use of safe math operands

* closing - without conviction

* better conviction recording

* ethparis conviction voting formula

* connect conviction voting pallet to runtime

* restore balance to u128

* accumulate conviction based on proposal age

* remove unused decay parameter

* make sure we log the creation block

* W3F Testing instructions

* Update types accordingly

* make formula accumulate over a longer time

* explain the conviction formula and state its limitations

+++ support/src/testing.rs
@@ -36,7 +36,7 @@ parameter_types! {
 
 pub mod primitives {
     pub type AccountId = u128;
-    pub type Balance = u64;
+    pub type Balance = u128;
     pub type CurrencyId = u8;
     pub type BlockNumber = u64;
 }
